requirements.txt
mlflow==2.20.1
numpy==2.0.2
pandas==2.2.3
scikit_learn==1.6.1
python-dotenv==1.0.1
nltk==3.9.1